% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMapMyCells.R
\name{addMapMyCells}
\alias{addMapMyCells}
\title{This function builds files needed for hierarchical mapping and stores them in the uns$hierarchical of AIT (Shiny) taxonomy.}
\usage{
addMapMyCells(
  AIT_anndata,
  hierarchy = AIT_anndata$uns$hierarchy,
  anndata_path = NULL,
  force = FALSE,
  n_processors = 3,
  normalization = "log2CPM",
  tmp_dir = NULL,
  user_precomp_stats_path = NULL,
  user_query_markers_path = NULL
)
}
\arguments{
\item{AIT_anndata}{A reference taxonomy anndata object.}

\item{hierarchy}{Named list of term_set_labels in the reference taxonomy ordered from most gross to most fine. Will default to list included in AIT_anndata, if any. E.g. \link{"Class" = 0, "Subclass"=  1}}

\item{anndata_path}{Local file path of the AIT reference taxonomy (h5ad file).}

\item{force}{Boolean value indicating whether to overwrite the AIT reference taxonomy's hierarchical file for a given mode.}

\item{n_processors}{Number of independent worker processes to spin up.}

\item{normalization}{Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.}

\item{tmp_dir}{Temporary directory for writing out the hierarchical files.}

\item{user_precomp_stats_path}{Alternative path to the user provided precompute stats HDF5 file. Will be generated, if not provided.}

\item{user_query_markers_path}{Alternative path to the user provided query markers JSON file. Will be generated, if not provided.}
}
\value{
AIT_anndata, a reference taxonomy with hierarchical files, such as precompute stats and query markers saved in uns.
}
\description{
This hierarchical mapping is a wrapper around cell_type_mapper and call's it's functions to generate needed files needed for mapping.
}
