% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchseqDirectory.R
\name{buildMappingDirectory}
\alias{buildMappingDirectory}
\title{Starting from an anndata object this function builds the minimum files required for patch-seq shiny}
\usage{
buildMappingDirectory(
  AIT.anndata,
  mappingFolder,
  query.data,
  query.metadata,
  query.mapping,
  doPatchseqQC = TRUE,
  metadata_names = NULL,
  min.confidence = 0.7,
  verbose = TRUE
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy object.}

\item{mappingFolder}{The location to save output files for patch-seq (or other query data) results, e.g. "/allen/programs/celltypes/workgroups/rnaseqanalysis/shiny/star/human/human_patchseq_MTG_JAM_TEST/current/"}

\item{query.data}{A CPM normalized matrix to be annotated.}

\item{query.metadata}{A data frame of metadata for the query data.}

\item{query.mapping}{Mapping results from \code{taxonomy_mapping()}, must be an mappingClass S4 object. If provided row names must match column names in query.data.}

\item{doPatchseqQC}{Boolean indicating whether patch-seq QC metrics should be calculated (default) or not.}

\item{metadata_names}{An optional named character vector where the vector NAMES correspond to columns in the metadata matrix and the vector VALUES correspond to how these metadata should be displayed in Shiny. This is used for writing the desc.feather file later.}

\item{min.confidence}{Probability below which a cell cannot be assigned to a cell type (default 0.7).  In other words, if no cell types have probabilities greater than resolution.index, then the assigned cluster will be an internal node of the dendrogram.}

\item{verbose}{Should detail logging be printed to the screen?

This function writes files to the mappingFolder directory for visualization with molgen-shiny tools
--- anno.feather - query metadata
--- data.feather - query data
--- dend.RData   - dendrogram (copied from reference)
--- desc.feather - table indicating which anno columns to share
--- memb.feather - tree mapping of each query cell to each tree node (not just the best matching type like in treeMap)
--- tsne.feather - low dimensional coordinates for data
--- tsne_desc.feather - table indicating which low-D representations to share}
}
\description{
Starting from an anndata object this function builds the minimum files required for patch-seq shiny
}
