% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTaxonomy.R
\name{checkMetadata}
\alias{checkMetadata}
\title{Checks whether metadata is in scrattch.taxonomy format}
\usage{
checkMetadata(
  meta.data,
  schema,
  messages = c(),
  isValid = FALSE,
  isWarning = FALSE,
  print.messages = TRUE,
  ...
)
}
\arguments{
\item{meta.data}{A meta.data object to be tested}

\item{schema}{The schema data.frame.}

\item{messages}{The current messages to append to.}

\item{isValid}{The current call for whether the taxonomy is valid.}

\item{isWarning}{The current call for whether the taxonomy is valid.}

\item{print.messages}{Print messages only to a log file (FALSE; default) or also to the screen (TRUE)}

\item{...}{Additional parameters for ._validate_schema_element and .validate_var_elements (can be ignored in most cases)}
}
\value{
Logical vector indicating whether the inputted taxonomy is a valid scrattch.taxonomy format.
}
\description{
Messages will be printed to the screen.
}
