% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTaxonomy.R
\name{checkTaxonomy}
\alias{checkTaxonomy}
\title{Checks whether an anndata object is in scrattch.taxonomy format and returns a log-file if not}
\usage{
checkTaxonomy(
  AIT.anndata,
  log.file.path = getwd(),
  print.messages = FALSE,
  pull_ensembl = FALSE,
  validate_percent_ensembl = 60,
  pull_assay = FALSE,
  pull_cl = FALSE,
  validate_percent_cl = 80,
  pull_ncbitaxon = FALSE,
  pull_uberon = FALSE,
  pull_brain_atlases = FALSE,
  pull_hancestro = FALSE,
  pull_mondo = FALSE,
  ...
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object to be tested}

\item{log.file.path}{The directory to output the logfile of errors and warnings (if any; default getwd())}

\item{print.messages}{Print messages only to a log file (FALSE; default) or also to the screen (TRUE)}

\item{pull_ensembl}{If FALSE (default) loads the list of Ensembl terms from NCBI; otherwise, pulls from NCBI (VERY slow)}

\item{validate_percent_ensembl}{Percent of entries that must correspond to valid Ensembl terms to validate}

\item{pull_assay}{If FALSE (default) loads the list of EFO terms (assays); otherwise pulls from EBI (VERY slow)}

\item{pull_cl}{If FALSE (default) loads a preset list of CL terms from OBO; otherwise pulls from OBO}

\item{validate_percent_cl}{Percent of entries that must correspond to valid CL terms to validate}

\item{pull_ncbitaxon}{If FALSE (default) loads the list of species with gene information at NCBI; otherwise pulls from OBO (VERY slow)}

\item{pull_uberon}{If FALSE (default) loads the list of anatomical regions from UBERON; otherwise pulls from OBO}

\item{pull_brain_atlases}{If FALSE (default) loads the list of brain atlas ids; otherwise pulls from brain-bican}

\item{pull_hancestro}{If FALSE (default) loads the list of HANCESTRO terms; otherwise, pulls from OBO}

\item{pull_mondo}{If FALSE (default) loads the list of MONDO terms; otherwise, pulls from OBO}

\item{...}{Additional parameters for ._validate_schema_elements and .validate_var_elements (can be ignored in most cases)}
}
\value{
Logical vector indicating whether the inputted taxonomy is a valid scrattch.taxonomy format.
}
\description{
Any breaking issues will cause this function to return FALSE.  And potential issues will still return TRUE but will output a warning to stderr.  All messages will get returned to the log file.
}
