% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createShiny.R
\name{createShiny}
\alias{createShiny}
\title{This function builds the minimum files required for Shiny}
\usage{
createShiny(AIT.anndata, shinyDir, metadata_names = NULL)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{shinyDir}{The location to save files for MolGen Shiny (AIBS)}

\item{metadata_names}{An optional named character vector where the vector NAMES correspond to columns in the metadata matrix and the vector VALUES correspond to how these metadata should be displayed in Shiny. This is used for writing the desc.feather file later.

Notes: Precomputed clusters must be provided.  In the anndata object these will be stored using the term "cluster".  If celltypeColumn is anything other than cluster, then any existing "cluster" column will be overwritten by celltypeColumn.  Values can be provided without colors and ids (e.g., "cluster") or with them (e.g., "cluster_label" + "cluster_color" + "cluster_id").  In this case cluster_colors is ignored and colors are taken directly from the metadata.  Cluster_id's will be overwritten to match dendrogram order.}
}
\value{
AIT anndata object in the specified format (only if return.anndata=TRUE)
}
\description{
This function builds the minimum files required for Shiny
}
