% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dend_to_json}
\alias{dend_to_json}
\title{Convert R dendrogram to json}
\usage{
dend_to_json(dend)
}
\arguments{
\item{dend}{R dendrogram object}
}
\value{
json file
}
\description{
Convert R dendrogram to json
}
