% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTaxonomyMetadata.R
\name{._find_best_ontology_match}
\alias{._find_best_ontology_match}
\title{This function will return the best matching ontology term id given a name and a distance to that best match}
\usage{
._find_best_ontology_match(search_term, ontology, top_n = 1)
}
\arguments{
\item{search_term}{Name of an ontology term you want to try and match (e.g., "frontal cortex" or "human")}

\item{ontology}{Variable of class ontology_index that holds a complete ontology read in from obo format using get_OBO.}

\item{top_n}{Number of top matches to return (default = 1)}
}
\value{
character indicating the associated id name and distance for the closest matching term
}
\description{
This function will return the best matching ontology term id given a name and a distance to that best match
}
\keyword{internal}
