% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTaxonomy.R
\name{._get_schema_def}
\alias{._get_schema_def}
\title{This function will return information about a given schema Key}
\usage{
._get_schema_def(key = NULL)
}
\arguments{
\item{key}{The schema Key to provide information about to the user.}
}
\value{
internal schema information
}
\description{
This function will return information about a given schema Key
}
\keyword{internal}
