% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTaxonomy.R
\name{._get_schema_elements}
\alias{._get_schema_elements}
\title{This function will return schema information for a given anndata Component and schema element Type}
\usage{
._get_schema_elements(schema, component, type = "MUST")
}
\arguments{
\item{schema}{The schema data.frame.}

\item{component}{The schema Component to provide information about.}

\item{type}{The schema element Type to provide information about.}
}
\value{
schema elements
}
\description{
This function will return schema information for a given anndata Component and schema element Type
}
\keyword{internal}
