% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTaxonomyMetadata.R
\name{._transfer_factor_levels}
\alias{._transfer_factor_levels}
\title{Convert one string back to a factor with matched, but different, values}
\usage{
._transfer_factor_levels(new.string, initial.factor)
}
\arguments{
\item{new.string}{A character string to convert to a factor}

\item{initial.factor}{A matched original factor variable to transfer to the string}
}
\value{
The character string with specific characters changed to lower case
}
\description{
This function is speficially targeting the changes.  When a vector gets minor updates due to misspellings or ontologies, this factor order from the original vector is retained, even though the values are slightly changed.
}
