% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firsttolower}
\alias{firsttolower}
\title{Convert the first character of a string to lowercase}
\usage{
firsttolower(str)
}
\arguments{
\item{str}{A character string}
}
\value{
The character string with specific characters changed to lower case
}
\description{
Convert the first character of a string to lowercase
}
