% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMapMyCells.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{This function looks for a valid hierarchy list.}
\usage{
get_hierarchy(AIT_anndata, hierarchy)
}
\arguments{
\item{AIT_anndata}{AIT reference taxonomy object.}

\item{hierarchy}{List of term_set_labels in the reference taxonomy ordered from most gross to most fine.}
}
\value{
Local file path to the AIT reference taxonomy h5ad file.
}
\description{
This function looks for a valid hierarchy list.
}
\keyword{internal}
