% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNCBITaxon}
\alias{getNCBITaxon}
\title{Determine the NCBITaxon ID for a species from the scientific name}
\usage{
getNCBITaxon(
  species,
  ncbitaxon_obo = NULL,
  obo.file = "ncbitaxon.obo",
  obo.url =
    "https://raw.githubusercontent.com/obophenotype/ncbitaxon/refs/heads/master/subsets/taxslim.obo"
)
}
\arguments{
\item{species}{Scientific name for one or more species. Not case sensitive, but make sure you spell it correctly or the function will return NA.}

\item{ncbitaxon_obo}{Variable in ontologyIndex format read in from obo.file file.}

\item{obo.file}{Location to look for or put the ncbitaxon.obo file, if ncbitaxon_obo is not provided}

\item{obo.url}{URL to download obo.file from, if not already downloaded. We recommend keeping the default and only switching to "http://purl.obolibrary.org/obo/ncbitaxon.obo" if absolutely necessary.}
}
\value{
an AIT.anndata object with the updated/additional vector of highly variable genes
}
\description{
This function returns the NCBITaxon ID for a species from the scientific name.  It's probably easier just looking it up via a google search, and most of the ones used at the Allen Institute are listed here: https://github.com/AllenInstitute/GeneOrthology.
}
