% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_to_dend}
\alias{json_to_dend}
\title{Convert json to R dendrogram}
\usage{
json_to_dend(json)
}
\arguments{
\item{json}{json from R dendrogram}
}
\value{
R dendrogram object
}
\description{
Convert json to R dendrogram
}
