% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2CPM_byRow}
\alias{log2CPM_byRow}
\title{Convert a matrix of raw counts to a matrix of log2(Counts per Million + 1) values}
\usage{
log2CPM_byRow(counts, sf = NULL, denom = 1e+06, offset = 1)
}
\arguments{
\item{counts}{A matrix, dgCMatrix, or dgTMatrix of count values}

\item{sf}{vector of numeric values representing the total number of reads. If count matrix includes all genes, value calulated by default (sf=NULL) will be accurate; however, if count matrix represents only a small fraction of genes, we recommend also providing this value.}

\item{denom}{Denominator that all counts will be scaled to. The default (1 million) is commonly used, but 10000 is also common for sparser droplet-based sequencing methods.}

\item{offset}{The constant offset to add to each cpm value prior to taking log2 (default = 1)}
}
\value{
a dgCMatrix of log2(CPM + 1) values
}
\description{
The input can be a base R matrix or a sparse matrix from the Matrix package.
}
\details{
This function expects that columns correspond to genes, and rows to samples by default and is equivalent to running logCPM with cells.as.rows=TRUE (but a bit faster).  By default the offset is 1, but to calculate just log2(counts per Million) set offset to 0.
}
