% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMapMyCells.R
\name{run_precomp_stats}
\alias{run_precomp_stats}
\title{Generates the precomputed_stats.h5 file, which is an HDF5 file that contains a serialization of the cell type taxonomy tree,
as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.}
\usage{
run_precomp_stats(
  anndata_path,
  n_processors,
  normalization,
  tmp_dir,
  hierarchy
)
}
\arguments{
\item{anndata_path}{Local file path of the AIT reference taxonomy (h5ad file).}

\item{n_processors}{Number of independent worker processes to spin up.}

\item{normalization}{Normalization of the h5ad files; must be either 'raw' or 'log2CPM'.}

\item{tmp_dir}{Temporary directory for writing out the hierarchical files (the code will clean these up after itself).}

\item{hierarchy}{List of term_set_labels in the reference taxonomy ordered from most gross to most fine.}
}
\value{
File path to the precompute stats file.
}
\description{
Generates the precomputed_stats.h5 file, which is an HDF5 file that contains a serialization of the cell type taxonomy tree,
as well as statistics about the reference cells that have been assigned to the cell types in the taxonomy.
}
\keyword{internal}
