% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupMapMyCells.R
\name{run_reference_markers}
\alias{run_reference_markers}
\title{Generates reference_marker.h5 file. The reference markers are, for every pair of leaf nodes in the cell type taxonomy tree,
every gene that could conceivably be a marker gene for discriminating between those two cell types.}
\usage{
run_reference_markers(precomp_stats_output_path, n_processors, temp_folder)
}
\arguments{
\item{precomp_stats_output_path}{Local file path to the generated or user provided precomputed_stats.h5 file.}

\item{n_processors}{Number of independent worker processes to spin up.}

\item{temp_folder}{Temporary directory for writing out the hierarchical files (the code will clean these up after itself).}
}
\value{
File path of the reference marker file.
}
\description{
Generates reference_marker.h5 file. The reference markers are, for every pair of leaf nodes in the cell type taxonomy tree,
every gene that could conceivably be a marker gene for discriminating between those two cell types.
}
\keyword{internal}
