% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsampleCells}
\alias{subsampleCells}
\title{Function to subsample cells}
\usage{
subsampleCells(cluster.names, subSamp = 25, seed = 5, use.historical = FALSE)
}
\arguments{
\item{cluster.names}{A vector of cluster names in the reference taxonomy.}

\item{subSamp}{Number of cells to keep per cluster.}

\item{seed}{Random seed used for subsampling.}

\item{use.historical}{Defualt (FALSE) uses new, faster implementation. Set to TRUE to use the historical, but slower implementation for back-compatibility.}
}
\value{
Boolean vector of cells to keep (TRUE) and cells to remove (FALSE)
}
\description{
Function to subsample cells
}
