% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tosentence}
\alias{tosentence}
\title{Convert to sentence case}
\usage{
tosentence(str)
}
\arguments{
\item{str}{A character string}
}
\value{
The character string in sentence case
}
\description{
Convert to sentence case
}
