% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateMarkerGenes}
\alias{updateMarkerGenes}
\title{Function to add or update marker genes for the current mode}
\usage{
updateMarkerGenes(
  AIT.anndata,
  marker.genes = NULL,
  list.name = paste0("marker_genes_", AIT.anndata$uns$mode)
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{list.name}{Which slot in obs to should the highly variable genes go?  Default is marker_genes_\link{mode}}

\item{variable.genes}{Set of variable genes to add to the list.name slot in obs.}
}
\value{
an AIT.anndata object with the updated/additional vector of marker genes

Note that this is a wrapper of updateHighlyVariableGenes with different default names
}
\description{
Function to add or update marker genes for the current mode
}
