% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTaxonomyMode.R
\name{buildTaxonomyMode}
\alias{buildTaxonomyMode}
\title{Save marker genes for patchSeqQC}
\usage{
buildTaxonomyMode(AIT.anndata, ...)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{...}{Additional variables to be passed to \code{addDendrogramMarkers}

The following variables are added to AIT.anndata$uns:
$dend[\link{mode.name}]
$filter[\link{mode.name}]
$QC_markers[\link{mode.name}]
...$markers,
...$countsQC,
...$cpmQC,
...$classBr,
...$subclassF,
...$allMarkers
$memb[\link{mode.name}]
...$memb.ref,
...$map.df.ref}

\item{subsample}{The number of cells to retain per cluster (default = 100).}

\item{subclass.column}{Column name corresponding to the moderate-resolution cell types used for the cell types of interest (default = "subclass_label").}

\item{class.column}{Column name corresponding to the low-resolution cell types used for the off-target cell types (default = "class_label").}

\item{off.target.types}{A character vector of off-target (also known as 'contamination') cell types.  This must include at least one of the cell types found in "class.column" and/or "subclass.column" (both columns are checked)}

\item{mode.name}{A name to identify the new taxonomy version.}

\item{subclass.subsample}{The number of cells to retain for PatchseqQC contamination calculation (default = 100, probably no need to change).}

\item{num.markers}{The maximum number of markers to calculate per node per direction (default = 50)}

\item{taxonomyDir}{The location to save shiny output (default = current working directory).}
}
\value{
AIT.anndata An updated AIT.anndata variable with the above content added to AIT.anndata$uns for the relevant mode.name.
}
\description{
This function saves all the variables required for applying the patchseq QC algorithm \code{pathseqtools} (which is an more flexible version of the \code{patchSeqQC} algorithm) to AIT.anndata$uns. This is only used for patch-seq analysis.  Requirements for input include:
}
