% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTaxonomyMetadata.R
\name{computeOntologyTerms}
\alias{computeOntologyTerms}
\title{Updates a metadata data frame with ontology terms}
\usage{
computeOntologyTerms(
  metadata,
  log.file.path = getwd(),
  log.file.name = "computeOntologyTerms_log.txt",
  standardize.metadata = FALSE,
  compute.ontology.terms = c("organism", "anatomical_region", "self_reported_sex",
    "self_reported_ethnicity", "assay", "disease"),
  compute.brain.atlas.terms = "DHBA",
  convert.regions.to.names = TRUE,
  compute.cl.terms = NULL
)
}
\value{
A list where "metadata" is the updated metadata file and there are additional list entries corresponding to conversions surrounding ontology terms.  The original metadata data frame is also returned.
}
\description{
See updateTaxonomyMetadata for details. computeOntologyTerms is a wrapper for updateTaxonomyMetadata with defaults to do compute the ontology terms for everything (except CL), but not do anything else.
}
