% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{._get_schema_def}
\alias{._get_schema_def}
\title{This function will print out infromation about a given schema Key}
\usage{
._get_schema_def(key = NULL)
}
\arguments{
\item{key}{The schema Key to provide information about to the user.}
}
\description{
This function will print out infromation about a given schema Key
}
