% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{loadTaxonomy}
\alias{loadTaxonomy}
\title{Read in a reference data set in Allen taxonomy format}
\usage{
loadTaxonomy(
  taxonomyDir,
  anndata_file = "AI_taxonomy.h5ad",
  log.file.path = getwd(),
  force = FALSE
)
}
\arguments{
\item{taxonomyDir}{Directory containing the Shiny taxonomy -OR- a direct h5ad file name.}

\item{anndata_file}{File name of the anndata object (.h5ad) to be loaded.}

\item{log.file.path}{Path to write log file to. Defaults to current working directory.}
}
\value{
Organized reference object ready for mapping against.
}
\description{
Read in a reference data set in Allen taxonomy format
}
