% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateHighlyVariableGenes}
\alias{updateHighlyVariableGenes}
\title{Function to update high variable genes for the current mode}
\usage{
updateHighlyVariableGenes(
  AIT.anndata,
  variable.genes = NULL,
  mode = AIT.anndata$uns$mode
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{variable.genes}{Set of variable genes used for correlation and Seurat mapping (and various other things). By default sets the new variable genes to match what is in AIT.anndata$var$highly_variable_genes, or returns the input starting AIT.anndata if it doesn't exist}

\item{mode}{Which mode to update variable genes for (defaults to current mode)}
}
\value{
an  AIT.anndata object with the updated high variable genes for the mode
}
\description{
Function to update high variable genes for the current mode
}
