% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateHighlyVariableGenes}
\alias{updateHighlyVariableGenes}
\title{Function to add or update highly variable genes for the current mode}
\usage{
updateHighlyVariableGenes(
  AIT.anndata,
  variable.genes = NULL,
  list.name = paste0("highly_variable_genes_", AIT.anndata$uns$mode)
  mode = AIT.anndata$uns$mode
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}
\item{variable.genes}{Set of variable genes to add to the list.name slot in obs. Can be provided as a logical or character vector.}

\item{list.name}{Which slot in obs to should the highly variable genes go?  Default is highly_variable_genes_\link{mode}}
}
\value{
an AIT.anndata object with the updated/additional vector of highly variable genes
}
\description{
Function to add or update highly variable genes for the current mode
\item{variable.genes}{Set of variable genes used for correlation and Seurat mapping (and various other things). By default sets the new variable genes to match what is in AIT.anndata$var$highly_variable_genes, or returns the input starting AIT.anndata if it doesn't exist}

\item{mode}{Which mode to update variable genes for (defaults to current mode)}
}
\value{
an  AIT.anndata object with the updated high variable genes for the mode
}
\description{
Function to update high variable genes for the current mode
}
